Introduce `RemoteConnectionConfig` and decrease mutation
`RemoteConnectionConfig`, similar to `rest.Config` is used to share
connection configuration amongst components. For now this only holds a
URL but it should be easy to extend that to hold moreconfiguration
(credentials, certs, ...) if need be.

We use this configuration to start those new components (Scheduler,
ControllerManager, VirtualKubelet). Those components have in common,
that they know how to start and stop itself, but it is very specific to
one runtime cycle to which APIServer they need to connect to. Thus this
configuration has been moved from being a member of the struct to and
parameter to the Start method.

Lastly, we also tried to minimize the mutation on the main structs of
the new components. This has two implications:
- For the templating of the arguments, we now use the
  `internal.ProcessState`. So certain available templating variables now
  have a different name, most notably: {Data,Cert,Conf}Dir -> Dir
- Some members of the structs are now set-once (they will not be changed
  after the defaulting), others will be overwritten after the defaulting
  -- and thus can be queried afterwards. The mutated bit on the new
  components is the URL only.

The plan ahead is to complitely get rid of any mutation on the main
structs but make a `RemoteConnectionConfig` available for each
component. Those could be returned from `Start()` and/or by an
additional method.
Also, in case that seems to be a proper way forward, chances are good,
that Etcd & APIServer will be refactored to implement the same
interface.