Add kubeadm-dind-cluster as submodule
The binary is found by the `BinPathFinder` which has been pulled out of
the `lightweight` package and has been adopted to work for multiple
packages.

Submoduling the dind implementation enables being specific about which
version of the dind-cluster.sh script to use. The path can still be
overwritten by the user of the framework, through setting appropriate
environment variables.