Safeguard ref dictionary access
Files are expected to have a license header - the exact format depends
on the specific type of file. When the file is executed through main,
the script ensures that only files with existing templates associated
with them are checked.

This commit adds similar security at the point where the comparison of
file vs. template takes place; to cover cases where that function is
called outside main.